Give the change-password form its own route on both clients
The form was a disclosure panel inside Settings on web and an inline block on
mobile. Moving it to /settings/password (web) and a pushed screen (mobile) is
what makes /.well-known/change-password possible: the W3C Change Password URL
is a redirect, so it needs an address that lands on the form itself. Safari's
Keychain, Chrome's Password Checkup and 1Password all open that path from a
compromised-password warning, and a collapsed panel on /settings would have
dropped the user on a settings list with nothing open.

Sub-routing Settings also brings it in line with the rest of the app, which
already nests: /workouts/new, /food/log, /weight/:id. Settings was the only
flat section on either platform.

The redirect is served three ways so it behaves identically wherever the app
runs: nginx for compose and for the Fly demo, and a vite middleware for dev
and preview. Without the vite one the SPA fallback would answer /.well-known
with index.html and React Router would bounce an unauthenticated visitor to
/login — and the e2e suite, which runs against the dev server, could not have
covered it.

Success now replaces the form rather than sitting above it. Three filled-looking
boxes under a success banner invite a second submit, which would fail, because
the current password is no longer current.

The Maestro flow's post-submit assertions moved with the screen: it asserted
"Preferences and account configuration" to prove the user had not been signed
out, and that string is no longer on the pushed screen.
