+++ fly/nginx.fly.conf
@@ -29,6 +29,15 @@ server {
         proxy_pass http://localhost:3000/health;
     }
 
+    # W3C Change Password URL. Password managers (Safari Keychain, Chrome Password
+    # Checkup, 1Password) open this path to take a user straight to the change-password
+    # form. Exact-match so it wins over the SPA fallback, which would otherwise answer
+    # with index.html and leave React Router to bounce an unauthenticated visitor to
+    # /login. The dev/preview equivalent lives in web/vite.config.ts.
+    location = /.well-known/change-password {
+        return 302 /settings/password;
+    }
+
     # SPA fallback
     location / {
         try_files $uri $uri/ /index.html;
